requirements.txt
torch
torchvision
matplotlib
numpy
pandas
Pillow
scikit-learn
scipy
tqdm
jupyter